ai-service/requirements.txt
akshare>=1.12.0
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
jieba>=0.42.1
snownlp>=0.12.3
plotly>=5.18.0
dashscope>=1.14.0
